pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "sentrybench"
version = "0.1.0"
description = "Defense-first backdoor evaluation harness and benchmarking toolkit"
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "SentryBench Team" }]
license = { file = "LICENSE" }
dependencies = [
  "PyYAML>=6.0",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]

[project.scripts]
sentrybench = "sentrybench.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
